Remove `Bootstrap`, `MinestomRootClassLoader`. Cleanup
