--- src/main/java/net/minestom/server/Bootstrap.java
@@ -1,25 +1,0 @@
-package net.minestom.server;
-
-import net.minestom.server.extras.selfmodification.MinestomRootClassLoader;
-
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-
-/**
- * Used to launch Minestom with the {@link MinestomRootClassLoader} to allow for self-modifications
- */
-public final class Bootstrap {
-
-    public static void bootstrap(String mainClassFullName, String[] args) {
-        try {
-            ClassLoader classLoader = MinestomRootClassLoader.getInstance();
-
-            Class<?> mainClass = classLoader.loadClass(mainClassFullName);
-            Method main = mainClass.getDeclaredMethod("main", String[].class);
-            main.setAccessible(true);
-            main.invoke(null, new Object[]{args});
-        } catch (InvocationTargetException | NoSuchMethodException | IllegalAccessException | ClassNotFoundException e) {
-            e.printStackTrace();
-        }
-    }
-}
