--- src/test/java/demo/Start.java
@@ -1,11 +1,0 @@
-package demo;
-
-import net.minestom.server.Bootstrap;
-
-public class Start {
-
-    public static void main(String[] args) {
-        Bootstrap.bootstrap("demo.Main", new String[0]);
-    }
-
-}
