--- src/test/java/improveextensions/mixins/InstanceContainerMixin.java
@@ -1,12 +1,0 @@
-package improveextensions.mixins;
-
-//@Mixin(InstanceContainer.class)
-public class InstanceContainerMixin {
-
-//    @Inject(method = "<init>", at = @At("RETURN"), require = 1)
-//    private void constructorHead(CallbackInfo ci) {
-//        System.out.println("Mixin into InstanceContainerMixin");
-//        MixinIntoMinestomCore.success = true;
-//    }
-
-}
