--- src/test/java/testextension/TestDemoLauncher.java
@@ -1,11 +1,0 @@
-package testextension;
-
-import net.minestom.server.Bootstrap;
-
-public class TestDemoLauncher {
-
-    public static void main(String[] args) {
-        Bootstrap.bootstrap("demo.Main", args);
-    }
-
-}
