+++ src/test/resources/extension.json
@@ -1,8 +1,4 @@
 {
   "entrypoint": "testextension.TestExtension",
-  "name": "Test_extension",
-  "codeModifiers": [
-    "testextension.TestModifier"
-  ],
-  "mixinConfig": "mixins.testextension.json"
+  "name": "Test_extension"
 }
