--- src/test/resources/improveextensions/disableearlyload/extension.json
@@ -1,6 +1,0 @@
-{
-  "entrypoint": "improveextensions.DisableEarlyLoad",
-  "name": "DisableEarlyLoad",
-  "codeModifiers": [],
-  "mixinConfig": "mixins.minestomcore.json"
-}
