--- src/test/resources/improveextensions/extension.json
@@ -1,6 +1,0 @@
-{
-  "entrypoint": "improveextensions.MixinIntoMinestomCore",
-  "name": "MixinIntoMinestomCore",
-  "codeModifiers": [],
-  "mixinConfig": "mixins.minestomcore.json"
-}
