--- src/test/resources/improveextensions/j9modules/extension.json
@@ -1,14 +1,0 @@
-{
-  "entrypoint": "improveextensions.MixinIntoMinestomCoreWithJava9ModuleOnClasspath",
-  "name": "MixinIntoMinestomCoreWithJava9ModuleOnClasspath",
-  "codeModifiers": [],
-  "mixinConfig": "mixins.minestomcore.json",
-  "externalDependencies": {
-    "repositories": [
-      {"name": "JCentral", "url": "https://jcenter.bintray.com/"}
-    ],
-    "artifacts": [
-      "org.mockito:mockito-core:2.28.2"
-    ]
-  }
-}
