--- src/test/resources/improveextensions/j9modules/mixins.minestomcore.json
@@ -1,10 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "improveextensions.mixins",
-  "target": "@env(DEFAULT)",
-  "compatibilityLevel": "JAVA_11",
-  "mixins": [
-    "InstanceContainerMixin"
-  ]
-}
