--- src/test/resources/improveextensions/missingmodifiers/extension.json
@@ -1,8 +1,0 @@
-{
-  "entrypoint": "improveextensions.missingmodifiers.MissingCodeModifiersExtension",
-  "name": "MissingCodeModifiersExtension",
-  "codeModifiers": [
-    "InvalidCodeModifierClass"
-  ],
-  "mixinConfig": "mixins.$invalid$.json"
-}
