--- src/test/resources/mixins.testextension.json
@@ -1,11 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "testextension.mixins",
-  "target": "@env(DEFAULT)",
-  "compatibilityLevel": "JAVA_11",
-  "mixins": [
-    "InstanceContainerMixin",
-    "DynamicChunkMixin"
-  ]
-}
